+++ src/lib.rs
@@ -22,7 +22,7 @@
 //!
 //! ```toml
 //! [dependencies.ftd2xx-embedded-hal]
-//! version = "~0.6.0"
+//! version = "~0.7.0"
 //! features = ["static"]
 //! ```
 //!
@@ -33,7 +33,7 @@
 //!
 //! ```toml
 //! [dependencies.ftd2xx-embedded-hal]
-//! version = "~0.6.0"
+//! version = "~0.7.0"
 //! ```
 //!
 //! # Examples
@@ -87,7 +87,7 @@
 //! [newAM/bme280-rs]: https://github.com/newAM/bme280-rs/blob/main/examples/ftdi.rs
 //! [udev rules]: https://github.com/newAM/libftd2xx-rs/#udev-rules
 //! [setup executable]: https://www.ftdichip.com/Drivers/CDM/CDM21228_Setup.zip
-#![doc(html_root_url = "https://docs.rs/ftd2xx-embedded-hal/0.6.0")]
+#![doc(html_root_url = "https://docs.rs/ftd2xx-embedded-hal/0.7.0")]
 #![forbid(missing_docs)]
 #![forbid(unsafe_code)]
 
@@ -101,14 +101,13 @@ mod spi;
 
 pub use delay::Delay;
 pub use gpio::OutputPin;
-pub use i2c::I2c;
-use libftd2xx::Ft2232h;
-use libftd2xx::Ft232h;
-use libftd2xx::Ft4232h;
-use libftd2xx::FtdiCommon;
+pub use i2c::{I2c, I2cError};
 pub use spi::Spi;
 
-use libftd2xx::{DeviceTypeError, Ftdi, FtdiMpsse, MpsseSettings, TimeoutError};
+use libftd2xx::{
+    DeviceTypeError, Ft2232h, Ft232h, Ft4232h, Ftdi, FtdiCommon, FtdiMpsse, MpsseSettings,
+    TimeoutError,
+};
 use std::convert::TryFrom;
 use std::{cell::RefCell, convert::TryInto, sync::Mutex, time::Duration};